fix: rename fuding_tests.rs to funding_tx_tests.rs
The test module file had a typo in its name ("fuding" instead of
"funding"), which was inconsistent with the module it tests.
